use textfield and dedupe insert
