+++ xmtp-inbox-ios/Localizable.strings
@@ -33,4 +33,5 @@
 "new-message" = "New message";
 "new-message-prompt" = "Enter a wallet address";
 "valid-ethereum-address" = "Valid Ethereum address";
+"invalid-ethereum-address" = "Invalid Ethereum address";
 "not-on-network-error" = "Recipient is not on the XMTP network";
